requirements.txt
docx==0.2.4
lxml==4.9.3
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
pydocstyle==6.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
snowballstemmer==2.2.0
tzdata==2023.3
